+++ examples/getstarted/src/api/address/routes/address.js
@@ -2,4 +2,11 @@
 
 const { createCoreRouter } = require('@strapi/strapi').factories;
 
-module.exports = createCoreRouter('api::address.address');
+module.exports = createCoreRouter('api::address.address', {
+  config: {
+    find: {
+      // auth: false,
+    },
+  },
+  only: ['find', 'findOne'],
+});
