+++ packages/plugins/graphql/package.json
@@ -61,7 +61,6 @@
     "graphql-depth-limit": "^1.1.0",
     "graphql-playground-middleware-koa": "^1.6.21",
     "graphql-scalars": "1.22.2",
-    "graphql-upload": "^15.0.0",
     "koa-bodyparser": "4.4.1",
     "koa-compose": "^4.1.0",
     "lodash": "4.17.21",
@@ -72,7 +71,6 @@
     "@strapi/strapi": "4.20.3",
     "@strapi/types": "4.20.3",
     "@types/graphql-depth-limit": "1.1.5",
-    "@types/graphql-upload": "15.0.2",
     "@types/koa-bodyparser": "4.3.12",
     "@types/koa__cors": "5.0.0",
     "cross-env": "^7.0.3",
