+++ packages/plugins/graphql/server/src/services/builders/queries/single-type.ts
@@ -45,8 +45,6 @@ export default ({ strapi }: Context) => {
     t: Nexus.blocks.ObjectDefinitionBlock<string>,
     contentType: Schema.SingleType
   ) => {
-    // const { uid } = contentType;
-
     const findQueryName = getFindOneQueryName(contentType);
     const typeName = getTypeName(contentType);
