+++ packages/plugins/graphql/server/src/services/constants.ts
@@ -1,6 +1,5 @@
 const PAGINATION_TYPE_NAME = 'Pagination';
 const DELETE_MUTATION_RESPONSE_TYPE_NAME = 'DeleteMutationResponse';
-const PUBLICATION_STATE_TYPE_NAME = 'PublicationState';
 const PUBLICATION_STATUS_TYPE_NAME = 'PublicationStatus';
 const ERROR_TYPE_NAME = 'Error';
 
@@ -156,7 +155,6 @@ export default () => ({
   PAGINATION_TYPE_NAME,
   RESPONSE_COLLECTION_META_TYPE_NAME,
   DELETE_MUTATION_RESPONSE_TYPE_NAME,
-  PUBLICATION_STATE_TYPE_NAME,
   PUBLICATION_STATUS_TYPE_NAME,
   GRAPHQL_SCALARS,
   STRAPI_SCALARS,
