+++ packages/plugins/graphql/server/src/services/internals/types/delete-mutation-response.ts
@@ -5,10 +5,6 @@ export default ({ strapi }: Context) => {
   const { DELETE_MUTATION_RESPONSE_TYPE_NAME } = strapi.plugin('graphql').service('constants');
 
   return {
-    /**
-     * Type definition for a Pagination object
-     * @type {NexusObjectTypeDef}
-     */
     DeleteMutationResponse: objectType({
       name: DELETE_MUTATION_RESPONSE_TYPE_NAME,
