+++ packages/plugins/graphql/server/src/services/internals/types/index.ts
@@ -1,7 +1,6 @@
 import pagination from './pagination';
 import buildResponseCollectionMeta from './response-collection-meta';
 import buildDeleteMutationResponse from './delete-mutation-response';
-import publicationState from './publication-state';
 import publicationStatus from './publication-status';
 import filters from './filters';
 import error from './error';
@@ -21,7 +20,6 @@ export default (context: Context) => () => {
     },
 
     [KINDS.enum]: {
-      publicationState: publicationState(context),
       publicationStatus: publicationStatus(context),
     },
